[Change] modify pipeline

+++ GBDT_add_nearby_data.py
@@ -101,6 +101,10 @@ def gbdt_add_nearby_stations_data(area, target_station, other_stations, hour, ta
 
     output_csv_path = '/'.join([data_root_folder, area, target_station, target_variable, hour, output_csv_name])
 
+    if os.exists(output_csv_path):
+        print('Nearby data already added. Passed. ' + output_csv_path)
+        return
+
     output = open(output_csv_path, "w+")
     wr = csv.writer(output)
